refactor(transformers): update downlevel ctor transformer
Follow the fix https://github.com/angular/angular/pull/40374
